bot/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
requests==2.32.0
firebase-admin==6.5.0
groq==0.13.0
